src/infra/db/mongodb/account-repository/account.ts
import { AddAccountRepository } from '../../../../data/protocols/add-account-repository'
import { AccountModel } from '../../../../domain/models/account-model'
import { AddAccountModel } from '../../../../domain/usecases/add-account'
import { MongoHelper } from '../helpers/mongo-helper'

export class AccountMongoRepository implements AddAccountRepository {
  async add (account: AddAccountModel): Promise<AccountModel> {
    const accountCollection = MongoHelper.getCollection('accounts')
    const result = await accountCollection.insertOne(account)
    const id = result.insertedId.toString()

    if (id) {
      return Object.assign({}, account, { id })
    }
    throw new Error('Falha na inserção do banco de dados')
  }
}
